requirements.txt
# Core Data Processing
pandas==2.1.0
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.0

# Database Connectivity
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pymongo==4.4.1
redis==4.6.0

# Data Validation & Quality
great-expectations==0.17.12
pandera==0.16.1
pydantic==2.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
bokeh==3.2.2

# ML & Statistics
statsmodels==0.14.0
imbalanced-learn==0.11.0
xgboost==1.7.6
category-encoders==2.6.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
joblib==1.3.2
pyarrow==13.0.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.7.0
flake8==6.1.0
mypy==1.5.1
pre-commit==3.3.3

# Logging & Monitoring
loguru==0.7.0
sentry-sdk==1.30.0

# API & Web (optional)
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0

# Documentation
sphinx==7.2.5
sphinx-rtd-theme==1.3.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.1
jupyterlab==4.0.5